requirements.txt
streamlit==0.61.0
Pillow==7.2.0
numpy
fastai==1.0.61
torch==1.6.0
torchvision==0.7.0
